+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Calendar Versioning](https://calver.org/#scheme) sc
 ----
 
 ## [Unreleased]
+
+## [2022.5.0-beta] - 2022/05/14
+_(dmidecode = 3.3 | smartmontools = 7.2 | hwinfo = 21.72 | lshw = 02.19 | lspci = 3.7.0)_
+
+- [fixed] timestamp generated in init
+- [added] settings decouple
 - [changed] name workbench lite to workbench core